requirements.txt
cmd2==2.4.3
colored==2.2.4
Jinja2==3.1.2
openpyxl==3.1.2
psycopg2_binary==2.9.9
PyPDF2==3.0.1
python_magic==0.4.27
SQLAlchemy==2.0.23
